Move to IAST tests
