+++ tracer/test/test-applications/security/Samples.Security.AspNetCore5/Controllers/IastController.cs
@@ -1164,8 +1164,6 @@ namespace Samples.Security.AspNetCore5.Controllers
         [FromQuery] string OutputType,
         [FromQuery] int SSRSReportID)
         {
-            // Get the value of the key the second item in Request.Query
-            var keyFirst = Request.Query.ElementAt(0).Key;
             var key = Request.Query.ElementAt(1).Key;
             var query1 = string.Format("\r\nDECLARE @{0}ID INT = (SELECT {0}ID FROM[Get{0}]('", key);
             var query2 = string.Format("'))\r\n\r\nSELECT SSRSReports FROM [ClientCentral].[dbo].[ClientDatabases] WHERE {0}ID = @{0}ID)", key);
